+++ gradle-test-extensions-subprojects/gradle-testkit-test-extensions/main/kotlin/com/javiersc/gradle/testkit/test/extensions/GradleTestKitExtensions.kt
@@ -168,4 +168,7 @@ public fun GradleRunner.andWithConfigurationCacheProblemsWarn() {
     withArguments(arguments + configurationCacheProblemsWarn())
 }
 
+public val BuildResult.outputTrimmed: String
+    get() = output.lines().joinToString("\n", transform = String::trim)
+
 private fun String.sanitizedSandboxPathPrefix(): String = replace("/", "--").replace("\\", "--")